[Game] GM script normalization (#1072)
- Changed most of the GM scripts output to use a standardized format (for V1 commands)
- Changed Faction Id from ``uint`` to ``FactionsEnum``
- Code cleaning on GM commands
- Removed obsolete chat functions
- Removed ``BroadcastPacketToFaction`` and ``BroadcastPacketToZone`` functions
- Removed ``/dummy`` and added a dummy option to use ``/spawn npc dummy`` instead
- Added debug area to ``/teleport``

+++ AAEmu.Game/Core/Packets/G2C/SCChatMessagePacket.cs
@@ -35,13 +35,13 @@ public class SCChatMessagePacket : GamePacket
     {
         stream.Write((short)_type);
         stream.Write((short)(_character?.Faction.Id ?? 0)); //chat
-        stream.Write(_character?.Faction.Id ?? 0); //chat, factionId?
+        stream.Write((uint)(_character?.Faction.Id ?? 0)); //chat, factionId?
         stream.WriteBc(_character?.ObjId ?? 0);
         stream.Write(_character?.Id ?? 0);
         stream.Write((byte)0); // TODO: remove this, use line under
         // stream.Write(_character != null ? _languageType : (byte) 0);
         stream.Write(_character != null ? (byte)_character.Race : (byte)0);
-        stream.Write(_character?.Faction.Id ?? 0); //type, factionId?
+        stream.Write((uint)(_character?.Faction.Id ?? 0)); //type, factionId?
         if (_character?.Connection?.GetAttribute("gmFlag") != null)
             stream.Write(_character != null ? "GM " + _character.Name : "");
         else
